prompt: add tests and refactor module
This adds some specs for the shell prompt module which was a bit
more involved than I originally though it would be. The main
thing is that I had to set up a temporary directory and do
some git commands. I also had to make sure not to print any
intermediate shell commands to stdout or stderr.