Add parser
This parses each line of input based on the tokens produced by the
tokenizer. It is necessary for both creating commands and checking
that the tokens are in a valid order. Tests have been added as well
to keep me honest. The command class was added in this commit too
since it's essentially the data class for the command.