Remove alias/command distinction
Aliases were not sufficiently distinct from commands so I've changed
things so that everything that the original idea of an alias
has been replaced by the command and is now called alias.

Before:
Alias - A command that is run without boolean logic (&&, || or ;)
Command - A command or a series of commands that can include boolean logic

After:
Alias - A command or a series of commands that can include boolean logic

It might make sense to revisit this at some point. Having variables
that store the output of commands and that can be used when creating
aliases would make a lot of sense here too. For example, you could
get the current branch name. The question is how often is this sort
of thing really needed or is it so rare that we can just omit it.